+++ api/python/src/MachO/objects/pyIndirectBindingInfo.cpp
@@ -1,0 +1,31 @@
+/* Copyright 2017 - 2024 R. Thomas
+ * Copyright 2017 - 2024 Quarkslab
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include "LIEF/MachO/IndirectBindingInfo.hpp"
+
+#include "MachO/pyMachO.hpp"
+
+namespace LIEF::MachO::py {
+
+template<>
+void create<IndirectBindingInfo>(nb::module_& m) {
+  nb::class_<IndirectBindingInfo, BindingInfo>(m, "IndirectBindingInfo",
+  R"delim(
+  This class represents a binding operation infered from the indirect symbol
+  table.
+  )delim"_doc);
+}
+
+}